requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-CORS==4.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.31.0
Werkzeug==3.1.3
hashlib==3.1.0
